+++ src/Symfony/Component/Notifier/Bridge/Infobip/InfobipTransport.php
@@ -85,7 +85,7 @@ final class InfobipTransport extends AbstractTransport
         return new SentMessage($message, (string) $this);
     }
 
-    protected function getEndpoint(): ?string
+    protected function getEndpoint(): string
     {
         return $this->host.($this->port ? ':'.$this->port : '');
     }