+++ src/Symfony/Component/Notifier/Transport/AbstractTransport.php
@@ -79,7 +79,7 @@ abstract class AbstractTransport implements TransportInterface
 
     abstract protected function doSend(MessageInterface $message): SentMessage;
 
-    protected function getEndpoint(): ?string
+    protected function getEndpoint(): string
     {
         return ($this->host ?: $this->getDefaultHost()).($this->port ? ':'.$this->port : '');
     }